requirements.txt
# Core Framework
Flask==3.0.2
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
PyMySQL==1.1.0
cryptography==42.0.5

# Auth
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1

# Validation
marshmallow==3.21.1
Flask-Marshmallow==1.2.0

# CORS
Flask-CORS==4.0.0

# Rate Limiting
Flask-Limiter==3.5.0

# AI (Groq - FREE)
groq==0.4.2

# PDF Generation
fpdf2==2.7.9

# Scheduler (reminders)
APScheduler==3.10.4

# Environment
python-dotenv==1.0.1

# HTTP
requests==2.31.0

# Development only
pytest==8.1.1
pytest-flask==1.3.0

# Deployment
gunicorn==22.0.0
